Import of ROOTKWOK/Win32-Process-Kill-v2.2 from CPAN.
gitpan-cpan-distribution: Win32-Process-Kill
gitpan-cpan-version:      v2.2
gitpan-cpan-path:         ROOTKWOK/Win32-Process-Kill-v2.2.zip
gitpan-cpan-author:       ROOTKWOK
gitpan-cpan-maturity:     released

+++ main.c
@@ -3,7 +3,7 @@
 typedef DWORD (*pNtTerminateProcess)(HANDLE,UINT);
 pNtTerminateProcess NtTerminateProcess;
 
-typedef int (*pRtlAdjustPrivilege)(int,BOOL,BOOL,int*);
+typedef LONG (*pRtlAdjustPrivilege)(int,BOOL,BOOL,int*);
 pRtlAdjustPrivilege RtlAdjPriv = NULL;
 
 typedef LONG (*pZwWriteVirtualMemory) (HANDLE ProcessHandle, PVOID BaseAddress, PVOID Buffer, ULONG BufferLength, PULONG ReturnLength);
